uuid is now generated upon import of the idf text file, as each JSON object inside an object (not array)  needs a unique identifier.  object_type is now pushed into the object alongside the Fields
